+++ domain/models/Drink.cs
@@ -4,7 +4,7 @@ namespace Models
     {
         public string DrinkType { get; set; }
 
-        public float Money { get; set; }
+        public double Money { get; set; }
 
         public int Sugars { get; set; }
